Changed file structure, changed entity search representation
Changed the representation of the entity search from a dictionary to a tree. This implementation is working nearly flawlessly, except that the "_in_question" attribute isn't cooperating and I can't figure out how to get it to show the proper question. Probably just because I'm tired...

I also changed the file structure of the project again because having a directory called "code" was interfering with PyCharm's debugger, and I care more about using that than having a directory named "code".